+++ sources/Application/Audio/DummyAudioOut.cpp
@@ -53,7 +53,6 @@ void DummyAudioOut::Stop() {
 	SAFE_DELETE(thread_) ;
 } ;
 
-void DummyAudioOut::SetSoftclip(int clip, int gain) {}
 void DummyAudioOut::SetMasterVolume(int volume) {}
 
 void DummyAudioOut::SendPulse()
