+++ sources/Services/Audio/AudioOutDriver.cpp
@@ -9,31 +9,9 @@
 #include <math.h>
 
 AudioOutDriver::AudioOutDriver(AudioDriver &driver) {
-    // Precalculate constant values for softclipping algorithm
-    softClipData_[0].alpha = 1.45f; // -1.5db (approx.)
-	softClipData_[1].alpha = 1.07f; // -3db (approx.)
-	softClipData_[2].alpha = 0.75f; // -6db (approx.)
-	softClipData_[3].alpha = 0.53f; // -9db (approx.)
-
-	for (int i = 0; i < 4; i++) {
-		softClipData_[i].alpha23 = softClipData_[i].alpha * (2.0f / 3.0f);
-		softClipData_[i].alphaInv = 1.0f / softClipData_[i].alpha;
-
-		if (softClipData_[i].alpha > 1.0f) {
-			/* calculates gain compensation differently for
-			 * modes with alpha > 1, so there's no drop in loudness
-			 * and we can still drive the hard clipper when the input
-			 * goes over 1.0
-			 */
-			softClipData_[i].gainCmp = 1.0f / (1.0f - (pow(softClipData_[i].alphaInv, 2.0f) / 3.0f));
-		} else {
-			softClipData_[i].gainCmp = 1.0f / softClipData_[i].alpha23;
-		}
-	}
-
-	driver_=&driver ;
-	driver.AddObserver(*this) ;
-	primarySoundBuffer_=0 ;
+    driver_ = &driver;
+    driver.AddObserver(*this) ;
+    primarySoundBuffer_=0 ;
     mixBuffer_ = 0;
     SetOwnership(false);
 }
@@ -56,7 +34,6 @@ void AudioOutDriver::Close() {
 }
 
 bool AudioOutDriver::Start() {
-    clipped_ = false;
     sampleCount_=0 ;
 	return driver_->Start() ;
 }
@@ -64,8 +41,6 @@ bool AudioOutDriver::Start() {
 void AudioOutDriver::Stop() {
 	driver_->Stop() ;
 }
-
-bool AudioOutDriver::Clipped() { return clipped_; };
 
 void AudioOutDriver::Trigger() {
 
@@ -84,58 +59,15 @@ void AudioOutDriver::Update(Observable &o,I_ObservableData *d)
 }
 
 void AudioOutDriver::prepareMixBuffers() {
-	sampleCount_=getPlaySampleCount() ; 
-	clipped_=false ;   
+    sampleCount_ = getPlaySampleCount();
 } ;
 
-void AudioOutDriver::SetSoftclip(int clip, int gain) {
-    softclip_ = clip - 1;
-	softclipGain_ = gain;
-}
-
 void AudioOutDriver::SetMasterVolume(int volume) {
-	masterVolume_ = volume;
-}
-
-fixed AudioOutDriver::hardClip(fixed sample) {
-
-    if (sample > MAX_POSITIVE_FIXED || sample < MAX_NEGATIVE_FIXED) {
-        clipped_ = true;
-		return sample > 0 ? MAX_POSITIVE_FIXED : MAX_NEGATIVE_FIXED;
-    }
-
-    return sample;
-}
-
-/* Implements standard cubic algorithm
- * https://wiki.analog.com/resources/tools-software/sigmastudio/toolbox/nonlinearprocessors/standardcubic
- */
-fixed AudioOutDriver::softClip(fixed sample) {
-    if (softclip_ == -1 || sample == 0)
-        return sample;
-
-    float x;
-    float sampleFloat = fp2fl(sample);
-	float maxFloat = fp2fl(sampleFloat > 0 ? MAX_POSITIVE_FIXED : MAX_NEGATIVE_FIXED);
-	SoftClipData* data = &softClipData_[softclip_];
-
-    x = data->alphaInv * (sampleFloat / maxFloat);
-    if (x > -1.0f && x < 1.0f) {
-        sampleFloat = maxFloat * (data->alpha * (x - (pow(x, 3.0f) / 3.0f)));
-    } else {
-        sampleFloat = maxFloat * data->alpha23;
-    }
-
-    if (softclipGain_) {
-        sampleFloat = sampleFloat * data->gainCmp;
-    }
-
-    return fl2fp(sampleFloat);
+    AudioMixer::SetMasterVolume(volume);
 }
 
 void AudioOutDriver::clipToMix() {
 
-    float damp = pow((float)masterVolume_ / 100, 4.0f);
     bool interlaced = driver_->Interlaced();
 
     if (!hasSound_) {
@@ -147,13 +79,10 @@ void AudioOutDriver::clipToMix() {
 
         fixed *p = primarySoundBuffer_;
 
-        fixed leftSample;
-        fixed rightSample;
-
         for (int i = 0; i < sampleCount_; i++) {
 
-            leftSample = damp * hardClip(softClip(*p++));
-            rightSample = damp * hardClip(softClip(*p++));
+            fixed leftSample = *p++;
+            fixed rightSample = *p++;
 
             *s1 = short(fp2i(leftSample));
             s1 += offset;
